requirements.txt
matplotlib==3.7.1
numpy==1.26.4
pillow==11.1.0
scikit-image==0.25.2
scikit-learn==1.5.1
seaborn==0.13.2
timm==1.0.7
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.44.1
